Split Host.Broker into crossplat and windows-specific parts: Part 2
