+++ src/Common/Core/Impl/OS/IProcess.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
+using System.IO;
+
+namespace Microsoft.Common.Core.OS {
+    public interface IProcess {
+        int Id { get; }
+        Stream StandardInput { get; }
+        Stream StandardOutput { get; }
+        Stream StandardError { get; }
+        bool HasExited { get; }
+        int ExitCode { get; }
+
+        event EventHandler Exited;
+
+        void Kill();
+    }
+}
