+++ src/Host/Broker/Impl/Resources.resx
@@ -117,6 +117,21 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="Critical_ParentProcessNotFound" xml:space="preserve">
+    <value>Designated parent process {0} not found</value>
+  </data>
+  <data name="Critical_PingTimeOut" xml:space="preserve">
+    <value>Ping timed out, terminating</value>
+  </data>
+  <data name="Default" xml:space="preserve">
+    <value>Default</value>
+  </data>
+  <data name="Error_BlockedByProfileDeletion" xml:space="preserve">
+    <value>User {0} session creation blocked due to pending profile deletion.</value>
+  </data>
+  <data name="Error_ExitRHost" xml:space="preserve">
+    <value>R Host Exited with Win32 error : {0}</value>
+  </data>
   <data name="Error_FailedRInstallationData" xml:space="preserve">
     <value>Failed to retrieve R installation data for interpreter \"{0}\" at \"{1}\"</value>
   </data>
@@ -129,11 +144,26 @@
   <data name="Error_ProfileDeletionFailedIO" xml:space="preserve">
     <value>Profile deletion failed for user {0}.</value>
   </data>
+  <data name="Error_RHostFailedToStart" xml:space="preserve">
+    <value>R session process failed to start. Error: {0}</value>
+  </data>
   <data name="Exception_PipeHasClientEnd" xml:space="preserve">
     <value>Pipe already has a client end</value>
   </data>
   <data name="Exception_PipeHasHostEnd" xml:space="preserve">
     <value>Pipe already has a host end</value>
+  </data>
+  <data name="Info_MonitoringParentProcess" xml:space="preserve">
+    <value>Monitoring parent process {0}</value>
+  </data>
+  <data name="Info_ParentProcessExited" xml:space="preserve">
+    <value>Parent process {0} exited, shutting down</value>
+  </data>
+  <data name="Info_StartingRHost" xml:space="preserve">
+    <value>Starting R session for connection {0} of user {1} with command line: {2}</value>
+  </data>
+  <data name="Info_StartedRHost" xml:space="preserve">
+    <value>R session started for connection {0} of user {1}</value>
   </data>
   <data name="Trace_CertificateIssuer" xml:space="preserve">
     <value>Certificate issued by {0}</value>
@@ -146,5 +176,8 @@
   </data>
   <data name="Trace_DetectedR" xml:space="preserve">
     <value>R {0} detected at \"{1}\"</value>
+  </data>
+  <data name="Trace_ErrorDataReceived" xml:space="preserve">
+    <value>|{0}|: {1}</value>
   </data>
 </root>
