+++ src/Host/Broker/Impl/Security/SecurityManager.cs
@@ -7,24 +7,23 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http.Authentication;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Microsoft.R.Host.Broker.Services;
 using Odachi.AspNetCore.Authentication.Basic;
 
 namespace Microsoft.R.Host.Broker.Security {
     public class SecurityManager {
         private readonly SecurityOptions _options;
-        private readonly ILogger _logger;
         private readonly IAuthenticationService _authenticationService;
 
-        public SecurityManager(IAuthenticationService authenticationService, IOptions<SecurityOptions> options, ILogger<SecurityManager> logger) {
+        public SecurityManager(IAuthenticationService authenticationService, IOptions<SecurityOptions> options) {
             _authenticationService = authenticationService;
             _options = options.Value;
-            _logger = logger;
         }
 
         public async Task SignInAsync(BasicSignInContext context) {
             ClaimsPrincipal principal;
             if (context.IsSignInRequired()) {
-                principal = _options.Secret != null ? SignInUsingSecret(context) : await _authenticationService.SignIn(context.Username, context.Password, context.Options.AuthenticationScheme);
+                principal = _options.Secret != null ? SignInUsingSecret(context) : await _authenticationService.SignInAsync(context.Username, context.Password, context.Options.AuthenticationScheme);
             } else {
                 var claims = new[] { new Claim(ClaimTypes.Anonymous, "") };
                 var claimsIdentity = new ClaimsIdentity(claims, context.Options.AuthenticationScheme);
