+++ src/Host/Broker/Impl/Services/ISystemInfoService.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace Microsoft.R.Host.Broker.Services {
+    public interface ISystemInfoService {
+        double GetCpuLoad();
+        (long TotalVirtualMemory, long FreeVirtualMemory, long TotalPhysicalMemory, long FreePhysicalMemory) GetMemoryInformation();
+        (string VideoCardName, long VideoRAM, string VideoProcessor) GetVideoControllerInformation();
+    }
+}
