+++ src/Host/Broker/Impl/Sessions/SessionStateChangedEventArgs.cs
