+++ src/Windows/Host/Broker/Impl/Resources.resx
@@ -123,12 +123,6 @@
   <data name="Critical_NoTlsCertificate" xml:space="preserve">
     <value>TLS certificate not found: {0}</value>
   </data>
-  <data name="Critical_ParentProcessNotFound" xml:space="preserve">
-    <value>Designated parent process {0} not found</value>
-  </data>
-  <data name="Critical_PingTimeOut" xml:space="preserve">
-    <value>Ping timed out, terminating</value>
-  </data>
   <data name="Critical_PipeConnectTimeOut" xml:space="preserve">
     <value>Requested to write server.urls to pipe '{0}', but timed out while trying to connect to pipe</value>
   </data>
@@ -144,9 +138,6 @@
   <data name="Error_ProfileCreationFailed" xml:space="preserve">
     <value>Profile creation failed for user {0} with WIN32 error code 0x{1}</value>
   </data>
-  <data name="Error_RHostFailedToStart" xml:space="preserve">
-    <value>R session process failed to start. Error: {0}</value>
-  </data>
   <data name="Error_UserNameParse" xml:space="preserve">
     <value>User name parsing failed for user {0} with WIN32 error code 0x{1}</value>
   </data>
@@ -156,23 +147,11 @@
   <data name="Info_BrokerName" xml:space="preserve">
     <value>Broker name '{0}' assigned</value>
   </data>
-  <data name="Info_MonitoringParentProcess" xml:space="preserve">
-    <value>Monitoring parent process {0}</value>
-  </data>
-  <data name="Info_ParentProcessExited" xml:space="preserve">
-    <value>Parent process {0} exited, shutting down</value>
-  </data>
   <data name="Info_ProfileAlreadyExists" xml:space="preserve">
     <value>Profile already exists for user: {0}</value>
   </data>
   <data name="Info_ProfileCreated" xml:space="preserve">
     <value>Profile created for user: {0}</value>
-  </data>
-  <data name="Info_StartedRHost" xml:space="preserve">
-    <value>R session started for connection {0} of user {1}</value>
-  </data>
-  <data name="Info_StartingRHost" xml:space="preserve">
-    <value>Starting R session for connection {0} of user {1} with command line: {2}</value>
   </data>
   <data name="Info_UserProfileServiceName" xml:space="preserve">
     <value>Microsoft.R.Host.UserProfile</value>
@@ -182,9 +161,6 @@
   </data>
   <data name="Trace_EnvironmentVariableCreationBegin" xml:space="preserve">
     <value>Creating user environment variables for user {0} with profile directory {1}</value>
-  </data>
-  <data name="Trace_ErrorDataReceived" xml:space="preserve">
-    <value>|{0}|: {1}</value>
   </data>
   <data name="Trace_LogOnSuccess" xml:space="preserve">
     <value>Log on succeeded for user: {0}</value>
@@ -210,20 +186,11 @@
   <data name="Error_ProfileCreationFailedInvalidResponse" xml:space="preserve">
     <value>Profile creation failed for user {0} with invalid response from the {1} service.</value>
   </data>
-  <data name="Error_BlockedByProfileDeletion" xml:space="preserve">
-    <value>User {0} session creation blocked due to pending profile deletion.</value>
-  </data>
-  <data name="Default" xml:space="preserve">
-    <value>Default</value>
-  </data>
   <data name="Text_ServiceDescription" xml:space="preserve">
     <value>Creates and manages R Host (Microsoft.R.Host.exe) sessions.</value>
   </data>
   <data name="Text_ServiceDisplayName" xml:space="preserve">
     <value>R Remote Service</value>
-  </data>
-  <data name="Error_ExitRHost" xml:space="preserve">
-    <value>R Host Exited with Win32 error : {0}</value>
   </data>
   <data name="Text_ServiceName" xml:space="preserve">
     <value>RRemoteService</value>
