+++ src/Windows/Host/Broker/Impl/project.json
@@ -15,7 +15,8 @@
     "Microsoft.Extensions.Logging.TraceSource": "1.1.1",
     "Microsoft.Extensions.Options.ConfigurationExtensions": "1.1.1",
     "Newtonsoft.Json": "9.0.1",
-    "System.Threading.Tasks.Dataflow": "4.5.24"
+    "System.Threading.Tasks.Dataflow": "4.5.24",
+    "System.ValueTuple": "4.3.0"
   },
   "frameworks": {
     "net46": {}
